+++ common/src/main/java/com/abrenoch/hyperiongrabber/common/HyperionScreenEncoderBase.java
@@ -1,23 +1,32 @@
 package com.abrenoch.hyperiongrabber.common;
 
+import android.content.res.Configuration;
 import android.media.projection.MediaProjection;
 import android.os.Handler;
 import android.os.HandlerThread;
 import android.util.Log;
 
 import com.abrenoch.hyperiongrabber.common.network.HyperionThread;
+import com.abrenoch.hyperiongrabber.common.util.BorderProcessor;
 import com.abrenoch.hyperiongrabber.common.util.HyperionGrabberOptions;
 
 public class HyperionScreenEncoderBase {
-    private static final boolean DEBUG = false;
+    static final boolean DEBUG = false;
     private static final String TAG = "ScreenEncoderBase";
-    int mWidthScaled;
-    int mFrameRate;
-    int mHeightScaled;
-    int mDensity;
-    Handler mHandler;
+    private final int CLEAR_COMMAND_DELAY_MS = 100;
+    private final int INIT_ORIENTATION;
 
+    final boolean mRemoveBorders = false; // enables detecting borders for standard grabbing - disabled for now
+    final boolean mAvgColor;
+    final int mWidthScaled;
+    final int mFrameRate;
+    final int mHeightScaled;
+    final int mDensity;
     boolean mIsCapturing = false;
+    int mCurrentOrientation;
+
+    BorderProcessor mBorderProcessor;
+    Handler mHandler;
     MediaProjection mMediaProjection;
     HyperionThread.HyperionThreadListener mListener;
 
@@ -30,12 +39,21 @@ public class HyperionScreenEncoderBase {
         mMediaProjection = projection;
         mDensity = density;
         mFrameRate = options.getFrameRate();
+        mAvgColor = options.useAverageColor();
+
+        int blackThreshold = options.getBlackThreshold();
+        mBorderProcessor = new BorderProcessor(blackThreshold);
+
+        mCurrentOrientation = INIT_ORIENTATION = width > height ? Configuration.ORIENTATION_LANDSCAPE :
+                Configuration.ORIENTATION_PORTRAIT;
 
         if (DEBUG) {
             Log.d(TAG, "Density: " + String.valueOf(mDensity));
             Log.d(TAG, "Frame Rate: " + String.valueOf(mFrameRate));
             Log.d(TAG, "Original Width: " + String.valueOf(width));
             Log.d(TAG, "Original Height: " + String.valueOf(height));
+            Log.d(TAG, "Average Color Only: " + String.valueOf(mAvgColor));
+            Log.d(TAG, "Black Pixel Threshold: " + String.valueOf(blackThreshold));
         }
 
         // find the common divisor for width & height best fit for the LED count (defined in options)
@@ -60,6 +78,12 @@ public class HyperionScreenEncoderBase {
 
     private Runnable clearAndDisconnectRunner = new Runnable() {
         public void run() {
+            if (DEBUG) Log.d(TAG, "Clearing LEDs and disconnecting");
+            try {
+                Thread.sleep(CLEAR_COMMAND_DELAY_MS);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
             mListener.clear();
             mListener.disconnect();
         }
@@ -67,6 +91,12 @@ public class HyperionScreenEncoderBase {
 
     private Runnable clearLightsRunner = new Runnable() {
         public void run() {
+            if (DEBUG) Log.d(TAG, "Clearing LEDs");
+            try {
+                Thread.sleep(CLEAR_COMMAND_DELAY_MS);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
             mListener.clear();
         }
     };
@@ -99,5 +129,17 @@ public class HyperionScreenEncoderBase {
 
     public void resumeRecording() {
         throw new RuntimeException("Stub!");
+    }
+
+    int getGrabberWidth() {
+        return INIT_ORIENTATION != mCurrentOrientation ? mHeightScaled : mWidthScaled;
+    }
+
+    int getGrabberHeight() {
+        return INIT_ORIENTATION != mCurrentOrientation ? mWidthScaled : mHeightScaled;
+    }
+
+    public void setOrientation(int orientation) {
+        mCurrentOrientation = orientation;
     }
 }