+++ CHANGELOG.md
@@ -1,4 +1,17 @@
-## [v0.4-alpha (unreleased)]
+## [v0.5-beta (unreleased)]
+### Changes
+- Added the ability to send only the average color of the screen
+- French translation
+- Norwegian translation
+- Czech translation
+- German translation
+- Dutch translation
+
+### Fixed
+- Lights now clear (if running) when shutting down
+- OOM bug
+
+## [v0.4-alpha]
 ### Changes
 - Start grabber on device boot
 - Added some eye candy for when grabber is started
@@ -7,13 +20,17 @@
 - New connection wizard
 - New settings/connection page (tv build)
 - Quick settings tile to toggle grabber (mobile build)
+- Screen orientation change updates grabber
+- Configurable grabber image quality
+- Pressing the notification will now return to the app's main activity
 
 ### Fixed
 - Grabber would fail to resume when waking device
 - OpenGL grabber sometimes halting immediately after starting screen grab
 - Default grabber failing to send data the first time it is turned on
-- Grabber not stoping when the host is unreachable
+- Grabber not stopping when the host is unreachable
 - Aspect ratio of grabbed image being slightly off
+- OOM bug
 
 ## [v0.3-alpha]
 ### Changes